Announce p2p swarm local&public IPs in docker ipfs
Currently when the dockerized ipfs was spinned up, it was only
announcing the docker network IPs in swarm, and therefore was not
accessible from outside. These changes will let the dockerized ipfs
instance announce the public IP + port, and local IP + port in swarm to
allow access from LAN and WAN.

The setting also uses
IPFS_LIBP2P_EXTERNAL_PORT that allows setting custom ports for the
dockerized ipfs node. This is useful for instance when there are
multiple ipfs nodes running on the same server and they need to listen to
different ports.