Fix for compilation buffer overflow (#543)
Fix the case when the output of the Solidity compiler is too large for
the stdout buffer. The fix changes the outputSelection field in the JSON
inputted to the compiler. The output now only contains byteCode.object
fields instead of the whole byteCode field, as other fields of the
output are not used at all.

Add error handling for buffer overflow or other compilation errors.

Change the metadata test of 'single.json' accordingly.

Renamed input variable to more explicit solcJsonInput.