Fix custom node-fetch in ipfs-util
The package ipfs-utils under ipfs-core is making use of a custom
node-fetch with version 2.6.7. The official node-fetch is latest at
2.6.6. Since the node-fetch versions were declared as node-fetch@^2.6.x
these are being overridden by the custom 2.6.7 version. This interferes
with builds on CircleCI.

Fix the node-fetch versions in the package to 2.6.6. I removed the
package-lock.json and ran a new npm install to make all packages use
2.6.6 except the ipfs-utils package.