Revert IPFS profile to server
On bf7bbab the profile of the server profile of the ipfs node
was removed to allow "local discovery" but it seems it is unrelated to
the networking issues at the time. Announcing the public ip's should
have been sufficient.

Instead the ipfs node started receiving the error:
"ignoring incoming dht message while not in server mode".

Revert back to the server profile.