Remove archiveWeb3,serve sourcify-chains /chains
The archiveWeb3 field under sourcify-chains.ts returns a full Web3
instance. This breaks the JSON format of the sourcify-chains if it were
to be served as a JSON. Plus, this field seems to never get used. This
is being mainly used for creation code fetching, but already all chains
with a  archiveWeb3 field have blockexplorer fetching, so this is never
used.

Finally, serve all chains of Sourcify (both supported:true and false)
under the endpoint /chains, as a single source of truth of Sourcify's
chains.