Enable badgerds, disable filestore
Enable badgerds for higher performance

Disable filestore since it changes the CIDs of added files. This is due
to filestore (--nocopy flag) using `raw` leaves instead of the default
`dag-pb`, which is also how the Solidity compiler calculates the ipfs
hashes. To be able serve the files with the same CIDs, we need to make
use of `dag-pb` too. This effects the real CID value while the
multiformat only chages how it is displayed (Qm.. vs. bafy...).