Fix eth-brownie output malformed metadata
Noticed by https://github.com/beamer-bridge/beamer which uses the brownie framework to compile contracts, that the output .json of brownie ir recognized wrongly as metadata file. This is because of the rather simplistic check when checking if a json is a metadata file.

Improve the metadata file check and malformed metadata message.