Fix partial match with bytecodeHash: none
Because of the check if the metadatahash exists, any contract without a
metadata hash i.e. with bytecodeHash: none set in the compiler settings
would receive a "partial match" with any arbitrary source code. There's
actually no need for such check since trimming the auxdata would suffice
for such a check.

Also rename trimMetadata trimAuxdata.

Thanks to @rimrakhimov for the heads up for this bug