Change conditional CI builds
The conditional CI setting was not always working as expected.
E.g. when there's a change in the server monitor was not building
or a change in the env vars wouldn't trigger a server build.

Monitor always depends on the server so whenever there's a change in
the server, build the monitor.

Also expand the conditions when server should be built.
The main reason the conditional built was put in place was to avoid
building IPFS each time and this causing a new IPFS add.
Unnecessary build times are still an issue but not a priority as the unnecessary
IPFS builds.

Mostly IPFS, UI, repo, and s3 are self-contained
and server/monitor more dependent on other parts of the project.
Keep their builds constrained, and expand server builds.

For this setting to work perfectly, each module needs to be
completely self contained.
I.e. Instead of server dependign on environments/.env values,
the values server needs has to be declared under src/server