Fix UI inputs
Fix some inputs being uncontrolled initially because the state values
are not initialized, ie. are undefined

Also fix constructor argument input fields not updating
because in setState, the prevState variable was being
mutated. This should not be mutated. Instead make a copy of the variable
and return that.

Fix form returning "Fill fields" error after typing once and deleting
fields.